processor can read props and react to props changes

+++ src/index.js
@@ -7,45 +7,67 @@ function connect(component,
                  processor) {
 
     const ImmviewConnector = React.createClass({
-
         componentWillMount() {
-            this.view = new View(sources, processor);
-            const cancelViewReaction = this.view.subscribe(() => this.forceUpdate());
+            this.view = new View(sources);
+            this.updateChildProps = (sourceData, props) => {
+                const childProps = {
+                    ...props,
+                    ...deImmubtablize(
+                        processor
+                            ? processor(sourceData, props)
+                            : sourceData
+                    ),
+                };
+                if (childProps) {
+                    this.setState({childProps});
+                }
+            };
+
+            const cancelViewReaction = this.view.subscribe(data => {
+                this.updateChildProps(data, this.props);
+            });
+
             this.destroyConnection = () => {
-                cancelViewReaction();
-                this.view.destroy();
+                if (this.view) {
+                    cancelViewReaction();
+                    this.view.destroy();
+                    this.view = null;
+                    this.destroyConnection = null;
+                    this.updateChildProps = null;
+                }
             };
         },
 
         componentWillUnmount() {
-            this.destroyConnection();
+            this.destroyConnection && this.destroyConnection();
+        },
+
+        componentWillReceiveProps(nextProps) {
+            this.updateChildProps(this.view.read(), nextProps);
         },
 
         render() {
-            const viewData = this.view.structure;
-            const viewProps = Iterable.isIterable(viewData) ? viewData.toObject() : viewData;
-            const props = {
-                ...this.props,
-                ...viewProps,
-            };
-            if (this.view.structure) {
-                return React.createElement(
-                    component,
-                    props,
-                    this.children
-                );
-            }
-
-            return null;
+            return React.createElement(
+                component,
+                this.state.childProps,
+                this.state.childProps.children
+            );
         },
 
     });
 
     return ImmviewConnector;
-
 }
 
-// for es6 import
-connect.default = connect;
+function deImmubtablize(d) {
+    return Iterable.isIterable(d) ? d.toObject() : d;
+}
 
+// for commonjs import
 module.exports = connect;
+
+// for es6 import
+module.exports.connect = connect;
+
+// for es6 default import
+module.exports.default = connect;
